requirements.txt
sentence-transformers
torch
scikit-learn
numpy
chromadb
google-generativeai
python-dotenv
ipykernel
streamlit
